move ray tracing structs to separate module
