+++ audionimbus/src/sealed.rs
@@ -1,0 +1,2 @@
+/// Sealed trait to prevent external implementations.
+pub trait Sealed {}
